miner/requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.6
idna==3.11
python-dotenv==1.2.2
requests==2.33.0
tree-sitter==0.25.2
tree-sitter-java==0.23.5
tree-sitter-python==0.25.0
urllib3==2.6.3
